Calibrate the vivax model

Uses the compartmental model to simulate the effect of interventions

calibrate_vivax_equilibrium(
  df,
  f = 1/72,
  gamma = 1/223,
  r = 1/60,
  delay = FALSE,
  rcd = FALSE,
  referral = FALSE,
  return.all = F,
  h.cutoff = 5e-08
)

Arguments

df

a dataframe containing the data, with one column called h containing the daily incidence and one variable called prop_import containing the proportion of imported cases among new infections. additional optional variables are alpha (effective care), beta (proportion of liver stage cure), rho (reporting rate) and omega (intensity of vector control)

f

relapse frequency

gamma

liver clearance rate

r

blood clearance rate

delay

a boolean indicating if the model including delays in treatment should be used. Default (FALSE) is the model without delay in treatment

rcd

a boolean indicating if the model including reactive case detection should be used. Default (FALSE) is the model without RCD

referral

a boolean indicating if the rcd model includes referral. Default (FALSE) is the model with referral for RCD. This parameter is used only if rcd==TRUE.

return.all

if TRUE, also returns delta and I estimates

h.cutoff

R0/Rc are not calculated for h values which are strictly inferior to h.cutoff

Value

A dataframe with 3 additional columns: lambda is the transmission rate, R0 is the basic reproduction number, Rc is the controlled reproduction number.

If lambda = -2, it means that the parameter combination does not correspond to a positive lambda solution with the model.

If lambda = -3, it means the incidence data contains missing values or infinite values or values below h.cutoff.

If return.all=T, delta and I are also included.

Details

If alpha is not provided in df, alpha=0. If beta is not provided in df, beta=1. If rho is not provided in df, rho=1. If omega is not provided in df, omega=1.