Simulate from equilibrium

Uses the compartmental model to simulate a trajectory, starting from equilibrium variable

simulate_from_equilibrium_fromdata(
  df,
  f = 1/72,
  gamma = 1/223,
  r = 1/60,
  maxtime,
  year
)

Arguments

df
a dataframe containing the data, with one column called I containing the proportion of infectious individuals (I0+Il) at equilibrium, one column called lambda containing the transmission rate, and one variable called id which identifies uniquely each row in the dataset. Additional optional variables are:
rho (reporting rate), delta (importation rate)
intervention levels in the past (when lambda was calculated): alpha.old (effective care), beta.old (proportion of liver stage cure), omega.old (intensity of vector control) intervention levels in the future (in the simulation): alpha.new (effective care), beta.new (proportion of liver stage cure), omega.new (intensity of vector control)

f
relapse frequency

gamma
liver clearance rate

r
blood clearance rate

maxtime
number of time steps for simulation

year
if TRUE, aggregates the outputs per year (h would be in cases per person year). if FALSE, returns daily outputs (h would be in cases per person day).

Value

A dataframe with the simulated state variables for each parameter combination in df

Details

If alpha is not provided in df, alpha=0. If beta is not provided in df, beta=1. If rho is not provided in df, rho=1. If omega is not provided in df, omega=1. If delta is not provided in df, delta=0.

Examples

mydata=data.frame(incidence=c(23,112),lambda=c(0.0063,0.0071),I=c(0.017,0.12),id=c(1,2))
mydata$rho=c(0.18,0.13)
mydata$beta.old=c(0.43,0.42)
mydata$alpha.old=c(0.17, 0.12)
mydata$delta=c(0,0)
mydata$omega.old=c(1,1)
simulate_from_equilibrium_fromdata(df=mydata, f=1/69, gamma=1/383, r=1/60,maxtime=2000,year=TRUE)

#> Warning: no omega.new in df, assumed omega.new=omega.old

#> Warning: no alpha.new in df, assumed alpha.new=alpha.old

#> Warning: no beta.new in df, assumed beta.new=beta.old

#>       time         Il          I0         Sl        S0          h         hr
#> 1        0 0.01471024 0.002289762 0.01615618 0.9668438 0.02230034 0.01538349
#> 366    365 0.01438409 0.002246759 0.01585660 0.9675126 0.02209045 0.01525157
#> 731    730 0.01409992 0.002202387 0.01554367 0.9681540 0.02165072 0.01494788
#> 1096  1095 0.01383094 0.002160387 0.01524746 0.9687612 0.02123283 0.01465798
#> 1461  1460 0.01357599 0.002120579 0.01496669 0.9693367 0.02083692 0.01438338
#> 1826  1825 0.01333404 0.002082798 0.01470021 0.9698829 0.02046135 0.01412293
#> 11       0 0.10443509 0.015564910 0.10578660 0.7742134 0.10832357 0.07274745
#> 3661   365 0.10611428 0.015772160 0.10717853 0.7709350 0.10904602 0.07320641
#> 7311   730 0.10727106 0.015943739 0.10833935 0.7684458 0.11035852 0.07412332
#> 10961 1095 0.10813078 0.016071248 0.10920197 0.7665960 0.11134152 0.07481433
#> 14611 1460 0.10876653 0.016165537 0.10983981 0.7652281 0.11206963 0.07532654
#> 18261 1825 0.10923493 0.016235005 0.11030972 0.7642203 0.11260672 0.07570460
#>                I p incidence id
#> 1     0.01700000 0  22.30034  1
#> 366   0.01663085 0  22.09045  1
#> 731   0.01630231 0  21.65072  1
#> 1096  0.01599133 0  21.23283  1
#> 1461  0.01569657 0  20.83692  1
#> 1826  0.01541684 0  20.46135  1
#> 11    0.12000000 0 108.32357  2
#> 3661  0.12188644 0 109.04602  2
#> 7311  0.12321480 0 110.35852  2
#> 10961 0.12420202 0 111.34152  2
#> 14611 0.12493207 0 112.06963  2
#> 18261 0.12546994 0 112.60672  2