Simulate the effect of several interventions

Uses the compartmental model to simulate the effect of interventions

simulate_vivax_interventions(
  df,
  intervention_list,
  f = 1/72,
  gamma = 1/223,
  r = 1/60,
  year = T,
  maxtime = 2000
)

Arguments

df
a dataframe containing the data, with one column called I containing the proportion of infectious individuals (I0+Il) at equilibrium, one column called lambda containing the transmission rate, and one variable called id which identifies uniquely each row in the dataset. Additional optional variables are:
rho (reporting rate), delta (importation rate)
intervention levels in the past (when lambda was calculated): alpha.old (effective care), beta.old (proportion of liver stage cure), omega.old (intensity of vector control) intervention levels in the future (in the simulation): alpha.new (effective care), beta.new (proportion of liver stage cure), omega.new (intensity of vector control) In practice, this dataframe can be the output of the function calculate_r0_rc_fromdata

intervention_list
a list of intervention objects.

f
relapse frequency

gamma
liver clearance rate

r
blood clearance rate

year
if TRUE, aggregates the outputs per year (h would be in cases per person year). if FALSE, returns daily outputs (h would be in cases per person day).

maxtime
number of time steps for simulation

Value

A dataframe with the simulated state variables for each parameter combination in df

Details

An intervention object is named list containing the intervention description. It should have the follwing structure: list(intervention_name="string", "alpha.new"=NA, "beta.new"=NA, "omega.new"=NA ) where NA can be replaced by scalars or kepts as such. If alpha.new is not provided in intervention_object it is equal to alpha. If beta.new is not provided in intervention_object it is equal to beta. If omega.new is not provided in intervention_object it is equal to omega.

Examples

mydata=data.frame(incidence=c(23,112),lambda=c(0.0063,0.0071),I=c(0.017,0.12),id=c(1,2))
mydata$rho=c(0.18,0.13)
mydata$beta.old=c(0.43,0.42)
mydata$alpha.old=c(0.17, 0.12)
mydata$delta=c(0,0)
mydata$omega.old=c(1,1)
int_0=list(intervention_name="0", "alpha.new"=NA, "beta.new"=NA, "omega.new"=NA )
int_A=list(intervention_name="A", "alpha.new"=NA, "beta.new"=0.6, "omega.new"=NA )
my_intervention_list=list(int_0, int_A)
simulate_vivax_interventions(df=mydata, my_intervention_list)

#> Warning: no alpha in df, assumed alpha=0

#> Warning: no beta in df, assumed beta=1

#> Warning: no omega in df, assumed omega=1

#> Warning: no alpha in df, assumed alpha=0

#> Warning: no beta in df, assumed beta=1

#> Warning: no omega in df, assumed omega=1

#>    id time          Il          I0          Sl        S0           h
#> 1   1    0 0.013413918 0.003586082 0.012097497 0.9709025 0.017955816
#> 2   1  365 0.011438304 0.003135727 0.010553559 0.9748724 0.016770936
#> 3   1  730 0.009903445 0.002714985 0.009137410 0.9782442 0.014504607
#> 4   1 1095 0.008610033 0.002360402 0.007944044 0.9810855 0.012593156
#> 5   1 1460 0.007511743 0.002059311 0.006930707 0.9834982 0.010974147
#> 6   1 1825 0.006573085 0.001801982 0.006064655 0.9855603 0.009593413
#> 7   1    0 0.013413918 0.003586082 0.012097497 0.9709025 0.017955816
#> 8   1  365 0.011438304 0.003135727 0.010553559 0.9748724 0.016770936
#> 9   1  730 0.009903445 0.002714985 0.009137410 0.9782442 0.014504607
#> 10  1 1095 0.008610033 0.002360402 0.007944044 0.9810855 0.012593156
#> 11  1 1460 0.007511743 0.002059311 0.006930707 0.9834982 0.010974147
#> 12  1 1825 0.006573085 0.001801982 0.006064655 0.9855603 0.009593413
#> 13  2    0 0.095543458 0.024456542 0.082828344 0.7971717 0.090162292
#> 14  2  365 0.078146883 0.020707634 0.069914061 0.8312314 0.082886255
#> 15  2  730 0.067640207 0.017923740 0.060514192 0.8539219 0.070860037
#> 16  2 1095 0.060261254 0.015968417 0.053912624 0.8698577 0.062563007
#> 17  2 1460 0.054815060 0.014525250 0.049040196 0.8816195 0.056536818
#> 18  2 1825 0.050646308 0.013420587 0.045310630 0.8906225 0.051978476
#> 19  2    0 0.095543458 0.024456542 0.082828344 0.7971717 0.090162292
#> 20  2  365 0.078146883 0.020707634 0.069914061 0.8312314 0.082886255
#> 21  2  730 0.067640207 0.017923740 0.060514192 0.8539219 0.070860037
#> 22  2 1095 0.060261254 0.015968417 0.053912624 0.8698577 0.062563007
#> 23  2 1460 0.054815060 0.014525250 0.049040196 0.8816195 0.056536818
#> 24  2 1825 0.050646308 0.013420587 0.045310630 0.8906225 0.051978476
#>             hr           I p incidence intervention
#> 1  0.011038966 0.017000000 0 17.955816            0
#> 2  0.010368869 0.014574032 0 16.770936            0
#> 3  0.008965132 0.012618430 0 14.504607            0
#> 4  0.007778262 0.010970435 0 12.593156            0
#> 5  0.006774277 0.009571054 0 10.974147            0
#> 6  0.005918990 0.008375067 0  9.593413            0
#> 7  0.011038966 0.017000000 0 17.955816            A
#> 8  0.010368869 0.014574032 0 16.770936            A
#> 9  0.008965132 0.012618430 0 14.504607            A
#> 10 0.007778262 0.010970435 0 12.593156            A
#> 11 0.006774277 0.009571054 0 10.974147            A
#> 12 0.005918990 0.008375067 0  9.593413            A
#> 13 0.054586180 0.120000000 0 90.162292            0
#> 14 0.050367855 0.098854517 0 82.886255            0
#> 15 0.042780824 0.085563947 0 70.860037            0
#> 16 0.037587687 0.076229671 0 62.563007            0
#> 17 0.033848732 0.069340310 0 56.536818            0
#> 18 0.031038390 0.064066895 0 51.978476            0
#> 19 0.054586180 0.120000000 0 90.162292            A
#> 20 0.050367855 0.098854517 0 82.886255            A
#> 21 0.042780824 0.085563947 0 70.860037            A
#> 22 0.037587687 0.076229671 0 62.563007            A
#> 23 0.033848732 0.069340310 0 56.536818            A
#> 24 0.031038390 0.064066895 0 51.978476            A