Transmission rate calculation

Solve the equation for lambda, vivax with CM, vector control, RCD and importation

solve_lambda_vivax_rcd(
  h,
  h1,
  r,
  f,
  gamma,
  alpha,
  beta,
  rho,
  p,
  omega,
  iota,
  nu,
  tau,
  eta,
  return.all = FALSE
)

Arguments

h

daily incidence

h1

daily incidence, directly detected cases

r

blood clearance rate

f

relapse frequency

gamma

liver clearance rate

alpha

effective treatment probability

beta

probability of liver clearance

rho

observation rate

p

proportion of imported cases

omega

vector control intensity. If vector control is not the purpose of the analysis, omega can be fixed to 1.

iota

RCD parameter: maximum number of index cases investigated per population per unit of time

nu

RCD parameter: number of secondary cases investigated per index case

tau

RCD parameter: targeting ratio of the RCD

eta

RCD parameter: probability that an investigated cases is detected (symptoms, test sensitivity, etc.)

return.all

if TRUE, returns lambda, I and delta. If FALSE, returns only lambda

Value

If return.all=TRUE, returns a list with lambda, I and delta. If return.all=FALSE, returns only a lambda value